feat:【PAC模板】流水线模板支持PAC特性 #11414 代码优化

+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/template/service/impl/TemplateInstallHistoryServiceImpl.kt
@@ -30,6 +30,7 @@ class TemplateInstallHistoryServiceImpl(
 
     override fun deleteVersions(srcTemplateCode: String, templateCode: String, versions: List<Long>) {
         logger.info("delete template install versions history:$srcTemplateCode|$templateCode|$versions")
+        if (versions.isEmpty()) return
         templateInstallHistoryDao.delete(
             dslContext = dslContext,
             srcTemplateCode = srcTemplateCode,
